LayoutTests/storage/indexeddb/getdatabases-expected.txt

Ensure all databases can be retrieved

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB.open('getdatabase-first')
PASS databases.length is 0
indexedDB.open('getdatabase-second', 2)
PASS databases.length is 1
PASS databases[0].version is 1
PASS databases[0].name is "getdatabase-first"
PASS databases.length is 2
PASS databases[0].version is 2
PASS databases[0].name is "getdatabase-second"
PASS successfullyParsed is true

TEST COMPLETE
